requirements.txt
opencv-python>=4.8.1
Pillow>=11.0.0
numpy>=2.0.0
tqdm>=4.66.1
